modev ip function to utils file
